+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 * Change add support for Python 3.14.0 and add to unittest matrix
 * Update apprise 1.9.2 (a2a2216) to 1.9.5 (19dfd14)
-* Update Beautiful Soup 4.13.4 (9752e85) to 4.13.5 (70c7473)
+* Update Beautiful Soup 4.13.4 (9752e85) to 4.14.2 (fc02060)
 * Update certifi 2025.08.03 to 2025.11.12
 * Update feedparser 6.0.11 (89fde03) to 6.0.12 (5d2578b)
 * Update idna library 3.10 (729225d) to 3.11 (ad949ee)
@@ -12,6 +12,7 @@
 
 [develop]
 * Update apprise 1.9.4 (1d9e1e5) to 1.9.5 (19dfd14)
+* Update Beautiful Soup 4.13.5 (70c7473) to 4.14.2 (fc02060)
 * Update certifi 2025.10.05 to 2025.11.12